Fix Limit and Filter params for cognito-idp list_users
